[sailfish-browser] Pass image contentType to DownloadManager. Fixes JB#10960
In the latest versions of gecko there is no single image cache, but one cache per
an open page (no to confuse with the permanent cache). In the same time there is
only one DownloadManager thus it cannot request contentType from a cache without
a relevant reference to the open web document. It's easier just to pass contentType
from ContextMenuHandler to DownloadManager.