[sailfish-browser] Update virtual keyboard visibility when browser page actived. Fixes JB#11991
This condition can be met when html page has a login form and
the user presses enter key from the keyboard to accept the form.
When enter is pressed, a dialog asking to save the credentials is brought on top of
the browser page. When dialog is closed, we need to verify that browser page triggers
visibility change for the input panel in case it was not yet done.