[sailfish-browser] Add new tab to the tab history. Fixes JB#14732
Link of the newly created tab was missing from the tab history (link table).
When user navigated to the third page from the initial
url of the tab, there was a previous page in the history. Now the
first url of the tab is treated as history entry as well.

This does not work entirely without declarativetab changes that
are coming in following commit.