[sailfish-browser] Remember zoom / scroll position for virtual tabs. Fixes JB#12738
Once maximum number of tabs (5) is reached, last tab is dropped as
virtual and css content rect is stored for the tab.

When a virtual tab is activated and content loaded, if you user has
not dragged, the content previous css content rect is restored.

This is not persisted to the database.

This commit also renames functions from TabModel that contain
"view" with "tab".