[sailfish-browser] Use engine's back/forward navigation when possible. Contributes to JB#10373
DeclarativeTab refactored to be more like a tab data container towards QML side and
it is not creatable from QML side. DeclarativeWebContainer provides currentTab
that can be used to access url, title, thumbnailPath of the current tab.

Back and forward navigation are going through DeclarativeWebContainer and it is
responsible for making decision if QuickMozView can be used for navigation.

Some refactoring are done to get this supported. Taking screen capture is
now implemented C++ side (DeclarativeWebContainer). No back/forward navigation
related code anymore in QML side.

There is an ugly meta object call to start loading when QuickMozView could not
be used for navigation. Currently, moving the WebView::load to C++ would also require
changes to the way how we deal with ConnectionHelper and I didn't want
to mix that into this commit.