Remove DeclarativeTab class
There are lot of small changes done. DeclarativeWebContainer
uses now only title and url from the active page and those
are also exposed to QML side. DeclarativeTabModel owns now one Tab
instance that matches to the current active tab. This is the tab that
is pushed back to model data when a new tab is activated or an active tab is changed.