[sailfish-browser] Fix page loading on the TabPage
In case active tab is closed from the TabPage and you have
multiple tabs, the next available tab is activated and loaded.
In above case, we need to be able to resetHeight for the page.

Normally, height is reset when loading starts and
when loading ends. When container is invisible
we should not go to the "boundHeightControl" state as it
blocks resetHeight.