[sailfish-browser] Load minimal amount of tabs immediately
Each fade animation is fixed now to animate only to the target
value and not animating always from zero. Now that tabs repeater
has the model always those animations were causing blinking when
entering directly to the history view. This also removes the tabs
grid not to blink when first time entered as opacity is already at
the target value.