[sailfish-browser] While activating a tab do not use page's title and url
While we are in process of activating a tab, do not pass title and url
of the active page through as it is already out dated. Previously this
caused an issue when cached url / title (m_url / m_title) got updated
and matching change signals were emitted right after new tab data became
available. When QML side tried to read updated value we returned
the value of the previous page.