[sailfish-browser] Add fader on top of the web content. Contributes to JB#29100
Fader is used to reduce flickering during orientation change. It fades
in with background color of the web page when rotation transition
starts. At the end of the orientation change transition pauses for 700ms
giving web engine a bit of time to relayout.

Content orientation is updated to the browser content window immediately
when fader has faded in.