[sailfish-browser] Refactor compositing surface clearing. Fixes JB#30162.
With the introduction of embedlite.compositor.request_external_gl_context_early
pref we can clear browser compositing surface directly from
DeclarativeWebContainer::createGLContext function. It should be now
called as soon as the view is created, not when the gecko engine paints
it's first frame.

Additionally this patch utilizes the QMozContext::PostCompositorTask API
to implement compositing surface clearing mechanims that works
regardless if some EmbedLite view has been created or not. Since both
the surface and the context are kept on the browser side we just need a
way to execute the clearing function on the correct thread.