[sailfish-browser] Add new tab cover action back to the tab view. Fixes JB#30443
This workaround adds the new tab cover action back to the cover in case
tab page is active when browser pushed to switcher. Other sub pages are not
handled (sharing and web prompts). For web prompts we should not have
new tab cover action as triggering the action would reject the prompt.

Remove this when the tab view cover is done.