[sailfish-browser] Trigger compositing surface clear earlier. Fixes JB#30162
Currently the assumption is that when we switch to live cover page the
some content is already rendered in the web window. Unfortunately gecko
triggers first repaint after initial page reflow has been done.
Depending on content complexity this can take several seconds. To
alleviate this problem we manually fill compositing surface with white
color on several ocasions. This currently happens when gecko requests GL
context from the platform (it happens when initial EmbedLitePuppetWidget
is created) and when we get expose event and there is no active page
which can render its content. In case of bug 30162 the clear operation
triggered from EmbedLitePuppetWidget creation should ensure that no
garbage is seen when switching to a live cover. Unfortunately the actual
EmbedLitePuppetWidget creation happens on a different thread than cover
switching. This leaves a tiny possible window in which garbage can be
seen. To fix this the patch adds additional clear action triggered
directly from cover switching code.