[sailfish-browser] Reset glScissor state before clering the surface. Fixes JB#30882
Gecko's CompositorOGL can use glScissor to limit the rendering area when
compositing pages containing inline surfaces (example: iframes). In case
we've closed a view which has set some scissor box the clear operation
will not actually clear the whole surface, but only part of it. The
remaining area will have undefined content. This patch ensures that we
always reset the scissor box before clearing the surface. We currenly
clear the surface manually only when no webview is opened so there is no
risk we'll affect operation of some existing CompositorOGL instance.
Ther shouldn't be any when the reset occours.