[sailfish-browser] Allow Homescreen to show browser cover when launching URLs and change favorite desktop configuration. Fixes JB#30997
Use the org.sailfishos.browser.ui service rather than
org.sailfishos.browser when opening URLs via libcontentaction. This
menas the Home screen will show the browser app cover while the app is
loading if triggered via an app-open action e.g. Qt.openUrlExternally().

However, browser bookmarks in the app grid are created using the "Link"
desktop entry, which already causes a separate desktop app to be
launched. To prevent two app covers from being displayed in in this
case, the .desktop configuration for grid bookmarks now uses the
"Application" entry type rather than "Link", with the URL passed via
the "Exec" entry. So, instead of launching a Link type that is
interpreted as a content action that should open the browser, the
favourite will now be launched directly with the browser app.