[sailfish-browser] Don't take focus away from the page when VKBD is closed. Fixes JB#31646.
The fact that virtual keyboard was closed for some reason does not imply
that the view is no longer focused. In the problematic case the fact
that we manually unfocused the view made the browser ignore subsequent
virtual keyboard show requests which originated from the engine. The
user is not the only source of focus changes, web pages can also modify
it's own focused element from JavaScript code.