[sailfish-browser] Improve tab thumbnailing logic. Fixes JB#23670
It's a bit of guess work but now we capture the third frame and
when the page is fully loaded. In addition, we do not clear the thumbnails
that aggresively as we are capturing only the active web page. Down
side is more or less same as it is now. For instance, if you close the
browser and leave it closed for a while, restart it, go to tab view and
you find that there out-dated thumbnails not matching the reality.

To get updates working properly, we refresh model data so (name is always
the same, like tab-1-thumb.jpg) that the new updated capture gets
rendered to the tab view.