[sailfish-browser] Prevent gecko from rendering when screen is off. Fixes JB#31866.
When the screen goes off we'd like to prevent gecko from doing any
rendering. We don't want to completely pause the compositor, but we also
don't want it to execute any GL/EGL commands, especially eglSwapBuffers.
The reason for this is because it seems that on sailfish eglSwapBuffers
can currently block indefinately when the display is turned off.

Blocked compositor thread was the main reason for huge memory leak
observed when the video was being played in the browser and the screen was
turned off. Gecko media thread produced new frame data which was never
consumed by the compositor thread. The data kept piling up until the
system run out of memory.