[sailfish-browser] Fix external link load when starting browser. Fixes JB#33892
New tab request was failing when browser was started so that there
tabs left from the previous browsing session.

Upon onNewTabRequest if iniailization cannot be completed save
the initial url and prepare a tab (activate). When the engine and all
facilities are ready, do not override the tab url within the
initialize.

The only special cases are that tab count is zero:
1) All tabs were closed on previous browsing session
2) "Private" => "History" cleanup was selected from Browser settings
   (browser was not running)

In case 1)
New tab requested is retrigger but no new pages created as
the next possible tab is already about to be activated. Thus,
we just wait.

In case 2)
If there were tabs from the previous browsing session and browser was
start from an external link (this doesn't actually matter but it is the most
complex scenario), all tabs are closed on initialize. Right after
all tabs are closed, a new tab is requested with the external link url.
As now the next possible tab is not yet active, a new tab will be
created from the external link.