[sailfish-browser] On dbus launch do not emit openUrlRequest signal
As external links are now opened to new tabs, do not trigger
openUrlRequest on startup if we're starting via dbus. Let the dbus
method take of requesting url. This was causing a bit of flickering
on the overlay (top and back to bottm) when starting the browser via
external url so that we did not have tabs.