[sailfish-browser] Set the default search engine after search engines have been added. Contributes to JB#32815
When browser is launched for the first time, the browser engine does not
know about the search engines. Once the browser engine is ready,
we request the engine to load the default configured search engines.

This commit changes browser initialization to listen "search-engine-added"
message. Once all initially configured engines are added, the default
search engine is set.

Browser does not handle yet a case where all search engines would
be changed on the fly.