[sailfish-browser] Do not wait for new tab when active tab is changed. Fixes JB#35786
When we are about to create a new tab we wait for a new tab. However,
when tab is switched or new tab is created it is not needed to
wait for a new tab.

This caused a glitch that occasionally loaded website was
still on the waiting mode (blank view) even thou it was
already loaded.