[sailfish-browser] Activate tab loading when overlay is dismissed. Fixes JB#35991
If active tab was closed from tabs view, new tab was triggered from
tab view, and cancelled from overlay by dragging the overlay, the active
tab id was not activated. The error did not occur, if overlay was
cancelled by clicking the mouse area above the overlay. This fixes
the error by activating tab always when overlay is dismissed.