[sailfish-browser] Do not change fullscreen mode value when chromeGestureEnabled changes. Fixes JB#36882
When an input field is focused, we do not want to update fullscreen mode.
This means that if content is on the fullscreen mode at the time of focusing
an input field, content remains on fullscreen mode. However, fullscreen mode
is updated if virtual keyboard is open and content is panned to the
beginning of the Y-axis. This guarantees that toolbar
will be appear after virtual keyboard is closed.

The fullscreenMode binding was originally introduced in commit sha1
f859aae593b65758fe (BrowserPage.qml) and back then we did not use margins
to drill hole for the virtual keyboard rather we changed height of
the web page. Thus, this is not needed and can be dropped.