[browser] Display SSL/TLS security info to the user. Contributes to JB#40089
Adds a new button to the toolbar that gives immediate feedback on the
SSL/TLS security status. Clicking on the button displays on overlay with
more info about the TLS certificate and security of the page being
viewed, as provided by gecko (via qtmozembed).

The certificate details page from jolla-settings-system is used to
display the certificate details to the user.