[sailfish-browser] Ensure empty browser opens with URL entry. Fixes JB#47875
When the browser opens for the first time there is no page to display,
and the user should be presented with the URL entry field open.

Previous changes which introduced the TLS panel broke this behaviour.
These changes fix the regression.

This change also sets the empty-page background colour to be the
Silica::overlayBackgroundColor.

The colour change makes the transition when opening the browser for the
first time, and when opening a new tab, less visually disruptive because
the page is the same colour as the overlay that appears over the top of
it. Otherwise, the screen flashes with the background colour before
being covered by the overlay.

The problem with the opening page was twofold. First, the atTop/atBottom
state behaviour changed (because panels with multiple sizes were
introduced, meaning it was possible to change from a large panel size to
a smaller panel size and still be atTop). The initial URL entry field
relies on a switch from atTop = false to atTop = true in order to
trigger a searchField.requestingFocus change in Overlay.qml. Second, the
initial _showUrlEntry entry also has to be true in order to trigger this
requestingFocus state change. This will revert to false as soon as
atBottom becomes true (i.e. as soon as the URL entry field is closed).