[browser] Delete startupCache if install is newer than it. Contributes to JB#35271
On installation of embedlite-components-qt5 the package installs the
file /var/lib/_MOZEMBED_CACHE_CLEAN_ .

This change ensures that when the browser is run, the last modified
timestamp of the startupCache is compared against this file, and if the
installed file is newer, the cache is cleared. The startupCache will
then be recreated by xulrunner during execution. Previously the
timestamp was compared against ~/.mozilla/mozembed/_CACHE_CLEAN_ but
since it's no longer createdi it can't be used anymore.

This ensures that any changes made to embedlite javascript components
won't be ignored in favour of older versions in the cache.

See JB#9389 for more info.