[sailfish-browser] Clean startupCache only using a oneshot. Contributes to JB#35271
The startupCache needs rebuilding if any of the components change, which
could happen if any one of xulrunner-qt5, embedlite-components-qt5 or
sailfish-browser is reinstalled.

Previously, the file /var/lib/_MOZEMBED_CACHE_CLEAN_ was touched on
installation and the modified date/time compared against the cache. This
change uses a oneshot, added by each of the components, instead.