[sailfish-browser] Introduce -captiveportal flag
It is going to be used to start a separate instance of browser
with a different profile folder for Wi-Fi captive portals.

Add a common class to retrieve profile name and if captive portal
mode is enabled.

Corresponding services and desktop file are added to start browser
from lipstick by changing D-Bus service name.

Contributes to JB#40777.