[ua] Updated user agent override for ebay properties. Contributes to JB#50424
This fixes two issues:

1. ebay login presents a blank page.
2. Selecting an ebay aution page causes a max-redirect error.

Ebay seems to have aggressive and broken user agent checks. The first
arose because of the inclusion of "Sailfish 3.0" in the UA string. The
second arose because of the inlusion of both "Mobile" and "Firefox",
which caused the desktop site (www.ebay.??) to be redirected to the
mobile site (m.ebay.??), which was then redirected to the desktop site,
and so on until the max redirect limit was hit.