[sailfish-browser] Make requestTab allowed for all. Fixes JB#52413
Privileged check doesn't work inside sandbox. There is little reason to
keep requestTab privileged API.

Allow closeTab for process that created the tab. It would be bad to
allow closeTab for all. Mark the process that called requestTab as owner
of the tab and allow the owner to call it.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>