[sailfish-browser] Add device lock gate for password access. Contributes to JB#54078 OMP#JOLLA-100
The Logins list in Settings allows access to the user's stored account
passwords, either by copying the password or editing it. This change
introduces a device lock gate, so that the user has to perform device
authentication (e.g. PIN, fingerprint) before the passwords will be
shown.

Once authenticated, the user won't be asked again until the Logins list
page has been exited.

The password access options are only available if an authentication
method is available.