[sailfish-browser] Avoid duplicating ConfigurationValue. JB#54316 OMP#JOLLA-150
If two or more ConfigurationValue instances are duplicated
simultaneously, the values are not correctly updated for sandboxed apps.
There is no problem if only a single instance exists at a time.

This change ensures the ConfigurationValue for the home page is shared
rather than duplicated when the folder picker is pushed.