[browser] Remove favicons when clearing history. Contributes to JB#54864 OMP#JOLLA-237
SettingManager::clearHistory() calls
DBManager::clearHistory() calls
DBWorker::clearHistory() emits historyAvailable() which triggers
DeclarativeHistoryModel::historyAvailable() which calls
DeclarativeHistoryModel::updateModel().

Thus, DeclarativeHistoryModel::clear() is never invoked in this
codepath.  We cannot simply clear the favicon for any url which is
no longer represented in the history model within updateModel()
since historyAvailable() will be emitted with limited data (due to
the LIMIT in the query in the general case, or due to the WHERE
in the dynamic filter case).  As such, we need to ensure that we
manually clear the history favicons when DBManager::clearHistory()
is called.