[browser] Do not reset model upon a tab close. Fixes JB#55345 OMP#JOLLA-333
We missed this when reviewing changes to the tab model, should not
have been added at the first place. This causes flickering in the
tab model when having multiple tabs open and closing one tab as
the model resets. If there are errors like model data not being
updated correctly, those should be fixed instead.