[saifish-browser] Use connectivity to establish online status. Fixes JB#55628 OMP#JOLLA-401
This partially reverts changes made in 8014ee097530c4c902. In that
change, ConnectionHelper was introduced to allow the "Select Wifi
network" dialog to be triggered. In order to simplify the code, the
online status check was also switched to using ConnectionHelper.

It turns out this interferes with the captive portal, since this relies
on the "ready" state to be considered equivalent to "online". The
change didn't respect this.

This change therefore ensures the browser thinks it's online as soon as
it's in the "Connected" state.