[browser] Remove SettingManager initialisation. Fixes JB#56200 OMP#JOLLA-507
The SettingManager initialisation was being skipped due to a small
regression from 68b52fef7fd89c12ddfe423.

The initialisation was needed to connect up a signal to the search
provider DConf value, without which any search provider changes made by
the user weren't enacted until the browser was restarted.

This change moves the signal connect into the constructor so that the
initialisation code is no longer needed.