[browser] Test requested external url against previous requested and resolved url. Fixes JB#56506 OMP#JOLLA-557
Previously requested external url is cleared upon real url change. This makes it possible to
find active tab in case user has not yet navigated on a tab that hold externally requested
url which may not be fully resolved.

Whenever user enters an url or it comes from an external source it's marked as external
and only after url is resolved it is updated to the database.

This commits adds requested url to the Tab struct, tst_persistenttabmodel unit test
case is changed to use Tab struct, and .