[browser] Quick copy url to clipboard by press-and-hold url from toolbar. Fixes JB#58623
This commit adds a way to quickly copy current url to clipboard
with long press. Before one would need to go via url text field
to copy the url to the clipboard. As all apps do not provide sharing
this simplifies sharing of the current url via clipboard.

Toolbar mouse area is changed to background item to take benefit of
cancelling of the action (among others). Platform defaults to 800ms
on long press action.

Signed-off-by: Raine Makelainen <raine.makelainen@jolla.com>