[browser] Delay external tab requests at startup. Fixes JB#57763
When the browser is started it loads the persistent tab model
asynchronously. Consequently if an external request to open a tab via
D-Bus triggered the browser to load (i.e. it wasn't already running),
the tab could be created before the tab model was loaded. This caused
the PID associated with the created tab to be wiped, and prevented the
external process that opened the tab from closing it again (used e.g. in
OAuth flows).

This change avoids this by delaying creation of tabs requested
externally until after the tab model has completed loading.