[browser] Remove D-Bus privilege checks
When the browser runs inside the sandbox it's not able to check whether
the sender is in the privileged group, so the check is guaranteed to
fail.

This change removes the check.

It leaves it in for the case of dumpMemoryInfo() as this may be useful
when running the browser unsandboxed for debugging purposes, and without
the check there would be no other restrictions on processes calling it.