[browser] Hide "hidden" windows
When a window has the "hidden" flag set, the window will immediately be
hidden after creation. In addition, such windows are set not to appear
in the tab view and are not saved persistently. As a consequence, such
windows aren't available for interaction by the user.

This is useful for printing, which needs a window to clone the page
content into during the print. This window appears as a blank window to
the user during the print. This change allows these windows to be
hidden.