[sailfish-browser] Further support for external flag
In some places a new tab is opened (e.g. when opening a bookmark or the captive
portal) but where the additional required fromExternal flag hadn't been added.
This prevented the calls to open a new tab from working.

This change adds the extra parameter to ensure the various functionalities
continue to work.

The flag is used to set the Sec-Fetch-* flags appropriately.